Renamed class and directories
