+++ dev/scripts/composer/ScriptInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace OpenMage\Scripts\Composer;
+
+use Composer\Script\Event;
+
+interface ScriptInterface
+{
+    public static function run(Event $event);
+}
